add Root component
