+++ src/index.js
@@ -5,13 +5,23 @@ import App from './components/App';
 import registerServiceWorker from './registerServiceWorker';
 import 'typeface-roboto';
 
-const rootEl = document.getElementById('root')
+import { Provider } from 'react-redux'
+import configureStore from './store'
 
-ReactDOM.render(<App />, rootEl);
+
+const rootEl = document.getElementById('root')
+const store = configureStore()
+
+ReactDOM.render(
+  <Provider store={store}>
+  <App />
+  </Provider>, rootEl);
 
 if (module.hot) {
   module.hot.accept('./components/App', () => {
-    ReactDOM.render(<App />, rootEl)
+    ReactDOM.render(<Provider store={store}>
+      <App />
+    </Provider>, rootEl)
   })
 }
