+++ src/reducers/chats.js
@@ -69,12 +69,13 @@ const byIds = (state = initialState.byIds, action) => {
          [getChatId(chat)]:chat
        })},{})
      }
+    case types.JOIN_CHAT_SUCCESS:
+    case types.LEAVE_CHAT_SUCCESS:
     case types.CREATE_CHAT_SUCCESS:
      return {
        ...state,
        [getChatId(action.payload.chat)]: action.payload.chat 
-     }
-    
+     }     
     case types.DELETE_CHAT_SUCCESS:
       const newState = { ...state }
       delete newState[getChatId(action.payload.chat)]
